requirements.txt
asgiref
Django
django-filter
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
PyJWT
pytz
sqlparse